refactor: 계층의 이름을 정확한 명칭으로 변경(#19)
