--- src/main/java/demago/gamjappang/user/application/port/in/JoinUseCase.java
@@ -1,7 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.JoinCommand;
-
-public interface JoinUseCase {
-    void join(JoinCommand command);
-}
