--- src/main/java/demago/gamjappang/user/application/port/in/LoginUseCase.java
@@ -1,8 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.LoginCommand;
-import demago.gamjappang.user.application.port.in.result.TokenResult;
-
-public interface LoginUseCase {
-    TokenResult login(LoginCommand command);
-}
