--- src/main/java/demago/gamjappang/user/application/port/in/LogoutUseCase.java
@@ -1,7 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.LogoutCommand;
-
-public interface LogoutUseCase {
-    void logout(LogoutCommand command);
-}
