--- src/main/java/demago/gamjappang/user/application/port/in/RefreshUseCase.java
@@ -1,8 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.RefreshCommand;
-import demago.gamjappang.user.application.port.in.result.TokenResult;
-
-public interface RefreshUseCase {
-    TokenResult refresh(RefreshCommand command);
-}
