--- src/main/java/demago/gamjappang/user/application/port/in/SendVerificationCodeUseCase.java
@@ -1,7 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.SendVerificationCodeCommand;
-
-public interface SendVerificationCodeUseCase {
-    void send(SendVerificationCodeCommand command);
-}
