--- src/main/java/demago/gamjappang/user/application/port/in/SocialLoginUseCase.java
@@ -1,8 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.SocialLoginCommand;
-import demago.gamjappang.user.application.port.in.result.TokenResult;
-
-public interface SocialLoginUseCase {
-    TokenResult socialLogin(SocialLoginCommand command);
-}
