--- src/main/java/demago/gamjappang/user/application/port/in/VerifyEmailUseCase.java
@@ -1,7 +1,0 @@
-package demago.gamjappang.user.application.port.in;
-
-import demago.gamjappang.user.application.port.in.command.VerifyEmailCommand;
-
-public interface VerifyEmailUseCase {
-    void verify(VerifyEmailCommand command);
-}
