--- src/main/java/demago/gamjappang/user/application/port/in/command/LoginCommand.java
@@ -1,4 +1,0 @@
-package demago.gamjappang.user.application.port.in.command;
-
-public record LoginCommand(String email, String password) {
-}
