--- src/main/java/demago/gamjappang/user/application/port/in/command/SendVerificationCodeCommand.java
@@ -1,4 +1,0 @@
-package demago.gamjappang.user.application.port.in.command;
-
-public record SendVerificationCodeCommand(String email) {
-}
