+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/JoinUseCase.java
@@ -1,0 +1,7 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.JoinCommand;
+
+public interface JoinUseCase {
+    void join(JoinCommand command);
+}
