+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/LoginUseCase.java
@@ -1,0 +1,8 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.LoginCommand;
+import demago.gamjappang.user.applicationcore.port.in.result.TokenResult;
+
+public interface LoginUseCase {
+    TokenResult login(LoginCommand command);
+}
