+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/LogoutUseCase.java
@@ -1,0 +1,7 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.LogoutCommand;
+
+public interface LogoutUseCase {
+    void logout(LogoutCommand command);
+}
