+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/RefreshUseCase.java
@@ -1,0 +1,8 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.RefreshCommand;
+import demago.gamjappang.user.applicationcore.port.in.result.TokenResult;
+
+public interface RefreshUseCase {
+    TokenResult refresh(RefreshCommand command);
+}
