+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/SendVerificationCodeUseCase.java
@@ -1,0 +1,7 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.SendVerificationCodeCommand;
+
+public interface SendVerificationCodeUseCase {
+    void send(SendVerificationCodeCommand command);
+}
