+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/SocialLoginUseCase.java
@@ -1,0 +1,8 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.SocialLoginCommand;
+import demago.gamjappang.user.applicationcore.port.in.result.TokenResult;
+
+public interface SocialLoginUseCase {
+    TokenResult socialLogin(SocialLoginCommand command);
+}
