+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/VerifyEmailUseCase.java
@@ -1,0 +1,7 @@
+package demago.gamjappang.user.applicationcore.port.in;
+
+import demago.gamjappang.user.applicationcore.port.in.command.VerifyEmailCommand;
+
+public interface VerifyEmailUseCase {
+    void verify(VerifyEmailCommand command);
+}
