+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/command/JoinCommand.java
@@ -1,4 +1,4 @@
-package demago.gamjappang.user.application.port.in.command;
+package demago.gamjappang.user.applicationcore.port.in.command;
 
 public record JoinCommand(String username, String email, String password) {
 }
