+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/command/LoginCommand.java
@@ -1,0 +1,4 @@
+package demago.gamjappang.user.applicationcore.port.in.command;
+
+public record LoginCommand(String email, String password) {
+}
