+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/command/LogoutCommand.java
@@ -1,4 +1,4 @@
-package demago.gamjappang.user.application.port.in.command;
+package demago.gamjappang.user.applicationcore.port.in.command;
 
 public record LogoutCommand(Long userId, String refreshToken) {
 }
