+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/command/RefreshCommand.java
@@ -1,0 +1,4 @@
+package demago.gamjappang.user.applicationcore.port.in.command;
+
+public record RefreshCommand(String refreshToken) {
+}
