+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/command/SendVerificationCodeCommand.java
@@ -1,0 +1,4 @@
+package demago.gamjappang.user.applicationcore.port.in.command;
+
+public record SendVerificationCodeCommand(String email) {
+}
