+++ src/main/java/demago/gamjappang/user/applicationcore/port/in/command/SocialLoginCommand.java
@@ -1,4 +1,4 @@
-package demago.gamjappang.user.application.port.in.command;
+package demago.gamjappang.user.applicationcore.port.in.command;
 
 public record SocialLoginCommand(String provider, String authorizationCode, String redirectUri) {
 }
